backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
httpx
discord.py
PyMySQL
authlib
